Made some minor performance improvements. Fixed head bobbing and is now more smooth. Game viewport now has an aspect ratio of 16:9

+++ game/ECS/Components/Camera.cs
@@ -3,6 +3,6 @@
     public struct Camera
     {
         public float fov;
-        public int bobFactor;
+        public float bobFactor;
     }
 }
